+++ src/pyannote/audio/core/model.py
@@ -24,6 +24,7 @@
 from __future__ import annotations
 
 import os
+import io
 import warnings
 from dataclasses import dataclass
 from functools import cached_property
@@ -51,6 +52,7 @@ from pyannote.audio.core.task import (
 from pyannote.audio.utils.hf_hub import AssetFileName, download_from_hf_hub
 from pyannote.audio.utils.multi_task import map_with_specifications
 from pyannote.audio.utils.dependencies import check_dependencies
+from pyannote.audio.telemetry import track_model_init
 
 
 # NOTE: needed to backward compatibility to load models trained before pyannote.audio 3.x
@@ -499,23 +501,25 @@ class Model(lightning.LightningModule):
     @classmethod
     def from_pretrained(
         cls,
-        checkpoint: Union[Path, Text],
+        checkpoint: Union[Path, Text, io.BytesIO],
         map_location=None,
         strict: bool = True,
         subfolder: Optional[str] = None,
         token: Union[str, bool, None] = None,
         cache_dir: Union[Path, str, None] = None,
+        skip_dependencies: bool = False,
         **kwargs,
     ) -> Optional["Model"]:
         """Load pretrained model
 
         Parameters
         ----------
-        checkpoint : Path or str
+        checkpoint : Path, str, or byte buffer
             Model checkpoint, provided as one of the following:
             * path to a local `pytorch_model.bin` model checkpoint
             * path to a local directory containing such a file
             * identifier of a model on Huggingface hub
+            * pre-loaded model checkpoint as a byte buffer
         map_location: optional
             Same role as in torch.load().
             Defaults to `lambda storage, loc: storage`.
@@ -531,6 +535,9 @@ class Model(lightning.LightningModule):
         kwargs: optional
             Any extra keyword args needed to init the model.
             Can also be used to override saved hyperparameter values.
+        skip_dependencies : bool, optional
+            If True, skip dependency check. Defaults to False.
+            Use at your own risk, as this may lead to unexpected behavior.
 
         Returns
         -------
@@ -542,24 +549,28 @@ class Model(lightning.LightningModule):
         `huggingface_hub.hf_hub_download`
 
         """
+        if isinstance(checkpoint, io.BytesIO) or os.path.isfile(checkpoint):
+            # if checkpoint is a BytesIO object or a file path, use it as is
+            path_to_model_checkpoint = checkpoint
+            otel_origin: str = "local"
+
+
         # if checkpoint is a directory, look for the model checkpoint
         # inside this directory (or inside a subfolder if specified)
-        if os.path.isdir(checkpoint):
+        elif os.path.isdir(checkpoint):
             if subfolder:
                 path_to_model_checkpoint = (
                     Path(checkpoint) / subfolder / AssetFileName.Model.value
                 )
             else:
                 path_to_model_checkpoint = Path(checkpoint) / AssetFileName.Model.value
-
-        # if checkpoint is a file, use it as is
-        elif os.path.isfile(checkpoint):
-            path_to_model_checkpoint = checkpoint
+            otel_origin: str = "local"
 
         # otherwise, assume that the checkpoint is hosted on HF model hub
         else:
+            checkpoint = str(checkpoint)
             _, _, path_to_model_checkpoint = download_from_hf_hub(
-                str(checkpoint),
+                checkpoint,
                 AssetFileName.Model,
                 subfolder=subfolder,
                 cache_dir=cache_dir,
@@ -567,6 +578,13 @@ class Model(lightning.LightningModule):
             )
             if path_to_model_checkpoint is None:
                 return None
+            
+            otel_origin: str = (
+                checkpoint
+                if checkpoint.lower().startswith(("pyannote/", "pyannoteai/"))
+                else "huggingface"
+            )
+
 
         if map_location is None:
 
@@ -578,9 +596,10 @@ class Model(lightning.LightningModule):
         # load checkpoint using lightning
         loaded_checkpoint = pl_load(path_to_model_checkpoint, map_location=map_location)
 
-        # check that the checkpoint is compatible with the current version
-        versions = loaded_checkpoint["pyannote.audio"]["versions"]
-        check_dependencies({"pyannote.audio": versions["pyannote.audio"]}, "Model")
+        if not skip_dependencies:
+            # check that the checkpoint is compatible with the current version
+            versions = loaded_checkpoint["pyannote.audio"]["versions"]
+            check_dependencies({"pyannote.audio": versions["pyannote.audio"]}, "Model")
 
         # obtain model class from the checkpoint
         module_name: str = loaded_checkpoint["pyannote.audio"]["architecture"]["module"]
@@ -589,6 +608,10 @@ class Model(lightning.LightningModule):
         Klass = getattr(module, class_name)
 
         try:
+            # if checkpoint is a BytesIO object, seek to the beginning so we can load it again
+            if isinstance(path_to_model_checkpoint, io.BytesIO):
+                path_to_model_checkpoint.seek(0)
+
             model = Klass.load_from_checkpoint(
                 path_to_model_checkpoint,
                 map_location=map_location,
@@ -597,6 +620,10 @@ class Model(lightning.LightningModule):
             )
         except RuntimeError as e:
             if "loss_func" in str(e):
+                # if checkpoint is a BytesIO object, seek to the beginning so we can load it again
+                if isinstance(path_to_model_checkpoint, io.BytesIO):
+                    path_to_model_checkpoint.seek(0)
+
                 msg = (
                     "Model has been trained with a task-dependent loss function. "
                     "Set 'strict' to False to load the model without its loss function "
@@ -612,5 +639,10 @@ class Model(lightning.LightningModule):
                 return model
 
             raise e
+
+        # save model origin (HF, local, etc) as attribute for telemetry purposes
+        model._otel_origin = otel_origin
+        track_model_init(model)
+
 
         return model