+++ src/pyannote/audio/pipelines/utils/getter.py
@@ -1,6 +1,7 @@
 # The MIT License (MIT)
 #
-# Copyright (c) 2017- CNRS
+# Copyright (c) 2017-2025 CNRS
+# Copyright (c) 2025- pyannoteAI
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -22,22 +23,53 @@
 
 import itertools
 from pathlib import Path
-from typing import Dict, Mapping, Optional, Text, Union
+from typing import Dict, Mapping
 
 import torch
+from pyannote.audio.core.calibration import Calibration
+from pyannote.audio.core.model import Model
+from pyannote.audio.core.pipeline import Pipeline
+from pyannote.audio.core.plda import PLDA
 from torch_audiomentations.core.transforms_interface import BaseWaveformTransform
 from torch_audiomentations.utils.config import from_dict as augmentation_from_dict
 
-from pyannote.audio.core.calibration import Calibration
-from pyannote.audio.core.model import Model
 
-PipelineModel = Union[Model, Text, Mapping]
+def get_pipeline(
+    pipeline: Pipeline | str | dict,
+    token: str | None = None,
+    cache_dir: Path | str | None = None,
+) -> Pipeline:
+    if isinstance(pipeline, Pipeline):
+        _pipeline = pipeline
+
+    elif isinstance(pipeline, str):
+        _pipeline = Pipeline.from_pretrained(pipeline, token=token, cache_dir=cache_dir)
+
+    elif isinstance(pipeline, dict):
+        pipeline.setdefault("token", token)
+        pipeline.setdefault("cache_dir", cache_dir)
+        _pipeline = Pipeline.from_pretrained(**pipeline)
+
+    else:
+        raise TypeError(
+            f"Unsupported type ({type(pipeline)}) for loading pipeline: "
+            f"expected `str` or `dict`."
+        )
+
+    if _pipeline is None:
+        raise ValueError(f"Could not load pipeline: {pipeline}.")
+
+    return _pipeline
+
+
+PipelineModel = Model | str | Mapping
 
 
 def get_model(
     model: PipelineModel,
-    token: Union[Text, None] = None,
-    cache_dir: Union[Path, Text, None] = None,
+    token: str | None = None,
+    cache_dir: Path | str | None = None,
+    skip_dependencies: bool = False,
 ) -> Model:
     """Load pretrained model and set it into `eval` mode.
 
@@ -52,6 +84,9 @@ def get_model(
         Huggingface token to be used for downloading from Huggingface hub.
     cache_dir: Path or str, optional
         Path to the folder where files downloaded from Huggingface hub are stored.
+    skip_dependencies : bool, optional
+        If True, skip dependency check. Defaults to False.
+        Use at your own risk, as this may lead to unexpected behavior.
 
     Returns
     -------
@@ -74,9 +109,13 @@ def get_model(
     if isinstance(model, Model):
         pass
 
-    elif isinstance(model, Text):
+    elif isinstance(model, str):
         _model = Model.from_pretrained(
-            model, token=token, cache_dir=cache_dir, strict=False
+            model,
+            token=token,
+            cache_dir=cache_dir,
+            strict=False,
+            skip_dependencies=skip_dependencies,
         )
         if _model:
             model = _model
@@ -84,6 +123,7 @@ def get_model(
     elif isinstance(model, Mapping):
         model.setdefault("token", token)
         model.setdefault("cache_dir", cache_dir)
+        model.setdefault("skip_dependencies", skip_dependencies)
         model = Model.from_pretrained(**model)
 
     else:
@@ -103,17 +143,18 @@ def get_model(
     return model
 
 
-PipelineAugmentation = Union[BaseWaveformTransform, Mapping]
+PipelineAugmentation = BaseWaveformTransform | Mapping
 
 
-PipelineCalibration = Union[Calibration, Text, Dict]
+PipelineCalibration = Calibration | str | Dict
 
 
 def get_calibration(
     calibration: PipelineCalibration,
-    token: Union[Text, None] = None,
-    cache_dir: Union[Path, Text, None] = None,
-) -> Optional[Calibration]:
+    token: str | None = None,
+    cache_dir: Path | str | None = None,
+    skip_dependencies: bool = False,
+) -> Calibration | None:
     """Load pretrained calibration
 
     Parameters
@@ -127,6 +168,9 @@ def get_calibration(
         Huggingface token to be used for downloading from Huggingface hub.
     cache_dir: Path or str, optional
         Path to the folder where files downloaded from Huggingface hub are stored.
+    skip_dependencies : bool, optional
+        If True, skip dependency check. Defaults to False.
+        Use at your own risk, as this may lead to unexpected behavior.
 
     Returns
     -------
@@ -141,14 +185,18 @@ def get_calibration(
     if isinstance(calibration, Calibration):
         loaded_calibration = calibration
 
-    elif isinstance(calibration, Text):
+    elif isinstance(calibration, str):
         loaded_calibration = Calibration.from_pretrained(
-            calibration, token=token, cache_dir=cache_dir
+            calibration,
+            token=token,
+            cache_dir=cache_dir,
+            skip_dependencies=skip_dependencies,
         )
 
     elif isinstance(calibration, Dict):
         calibration.setdefault("token", token)
         calibration.setdefault("cache_dir", cache_dir)
+        calibration.setdefault("skip_dependencies", skip_dependencies)
         loaded_calibration = Calibration.from_pretrained(**calibration)
 
     else:
@@ -158,6 +206,65 @@ def get_calibration(
         )
 
     return loaded_calibration
+
+
+PipelinePLDA = PLDA | str | Dict
+
+
+def get_plda(
+    plda: PipelinePLDA,
+    token: str | None = None,
+    cache_dir: Path | str | None = None,
+    skip_dependencies: bool = False,
+) -> PLDA | None:
+    """Load pretrained calibration
+
+    Parameters
+    ----------
+    plda : PLDA, str, or dict
+        When `PLDA`, returns `plda` as is.
+        When `str`, assumes that this is either the path to a checkpoint or the name of a
+        pretrained PLDA on Huggingface.co and loads with `PLDA.from_pretrained(PLDA)`.
+        When `dict`, loads with `PLDA.from_pretrained(**plda)`.
+    token : str or bool, optional
+        Huggingface token to be used for downloading from Huggingface hub.
+    cache_dir: Path or str, optional
+        Path to the folder where files downloaded from Huggingface hub are stored.
+    skip_dependencies : bool, optional
+        If True, skip dependency check. Defaults to False.
+        Use at your own risk, as this may lead to unexpected behavior.
+
+    Returns
+    -------
+    plda : PLDA
+        PLDA.
+
+    See also
+    --------
+    pyannote.audio.core.plda.PLDA.from_pretrained
+    """
+
+    if isinstance(plda, PLDA):
+        loaded_plda = plda
+
+    elif isinstance(plda, str):
+        loaded_plda = PLDA.from_pretrained(
+            plda, token=token, cache_dir=cache_dir, skip_dependencies=skip_dependencies
+        )
+
+    elif isinstance(plda, Dict):
+        plda.setdefault("token", token)
+        plda.setdefault("cache_dir", cache_dir)
+        plda.setdefault("skip_dependencies", skip_dependencies)
+        loaded_plda = PLDA.from_pretrained(**plda)
+
+    else:
+        raise TypeError(
+            f"Unsupported type ({type(plda)}) for loading PLDA: "
+            f"expected `str` or `dict`."
+        )
+
+    return loaded_plda
 
 
 def get_augmentation(augmentation: PipelineAugmentation) -> BaseWaveformTransform:
@@ -190,7 +297,7 @@ def get_augmentation(augmentation: PipelineAugmentation) -> BaseWaveformTransfor
     )
 
 
-def get_devices(needs: Optional[int] = None):
+def get_devices(needs: int | None = None):
     """Get devices that can be used by the pipeline
 
     Parameters